refactor(core-manager): move database-service.ts to database subfolder
